+++ src/integrators.rs
@@ -1,9 +1,0 @@
-pub mod verlet;
-pub mod runge_kutta;
-
-
-#[allow(dead_code)]
-pub enum SupportedIntegrator {
-    Verlet,
-    RungeKutta,
-}