+++ Changelog.md
@@ -1,3 +1,9 @@
+# Version 0.0.2
+
+- System with FCC lattice and Simple Cubic lattice.
+- Added ottimizations on the computation of potential and force: computation skipped if the distance of two atoms is greater than the cutoff.
+- Added a CLI commands to take the parameters from command line: only the *new run* is working.
+
 # Version 0.0.1
 
 Basic features: compute evolution of a system with FCC lattice, Verlet integrator and no optimizations.